colorbar combobox
