+++ napari/_qt/layer_controls/qt_image_controls_base.py
@@ -2,23 +2,14 @@ from contextlib import suppress
 from functools import partial
 
 import numpy as np
-from qtpy.QtCore import QModelIndex, QRect, Qt
-from qtpy.QtGui import QImage, QPainter, QPen, QPixmap
-from qtpy.QtWidgets import (
-    QComboBox,
-    QLabel,
-    QListView,
-    QPushButton,
-    QSlider,
-    QStyle,
-    QStyledItemDelegate,
-    QStyleOptionViewItem,
-)
+from qtpy.QtCore import Qt
+from qtpy.QtGui import QImage, QPixmap
+from qtpy.QtWidgets import QLabel, QPushButton, QSlider
 
-from ...utils.colormaps import ensure_colormap, make_colorbar
 from ..utils import qt_signals_blocked
 from ..widgets.qt_range_slider import QHRangeSlider
 from ..widgets.qt_range_slider_popup import QRangeSliderPopup
+from .qt_colormap_combobox import QtColormapComboBox
 from .qt_layer_controls_base import QtLayerControls
 
 
@@ -57,7 +48,7 @@ class QtBaseImageControls(QtLayerControls):
         self.layer.events.gamma.connect(self.gamma_slider_update)
         self.layer.events.contrast_limits.connect(self._on_clims_change)
 
-        comboBox = ColormapComboBox(self)
+        comboBox = QtColormapComboBox(self)
         comboBox.setObjectName("colormapComboBox")
         comboBox.addItems(self.layer.colormaps)
         comboBox._allitems = set(self.layer.colormaps)
@@ -306,69 +297,3 @@ def create_clim_reset_buttons(layer):
         range_btn.clicked.connect(reset_range)
 
     return reset_btn, range_btn
-
-
-COLORMAP_WIDTH = 150
-TEXT_WIDTH = 100
-ENTRY_HEIGHT = 30
-
-
-class ColorStyledDelegate(QStyledItemDelegate):
-    """
-    Class for paint :py:class:`~.ColorComboBox` elements when list trigger
-
-    :param base_height: height of single list element
-    :param color_dict: Dict mapping name to colors
-    """
-
-    def __init__(self, base_height: int, **kwargs):
-        super().__init__(**kwargs)
-        self.base_height = base_height
-
-    def paint(
-        self,
-        painter: QPainter,
-        style: QStyleOptionViewItem,
-        model: QModelIndex,
-    ):
-        rect = QRect(
-            style.rect.x(),
-            style.rect.y() + 2,
-            style.rect.width() - TEXT_WIDTH,
-            style.rect.height() - 4,
-        )
-        rect2 = QRect(
-            style.rect.width() - TEXT_WIDTH + 10,
-            style.rect.y() + 2,
-            style.rect.width(),
-            style.rect.height() - 4,
-        )
-        cbar = make_colorbar(ensure_colormap(model.data()), (18, 100))
-        image = QImage(
-            cbar, cbar.shape[1], cbar.shape[0], QImage.Format_RGBA8888,
-        )
-        painter.drawImage(rect, image)
-        painter.drawText(rect2, Qt.AlignCenter & Qt.AlignVCenter, model.data())
-        if int(style.state & QStyle.State_HasFocus):
-            painter.save()
-            pen = QPen()
-            pen.setWidth(5)
-            painter.setPen(pen)
-            painter.drawRect(rect)
-            painter.restore()
-
-    def sizeHint(self, style: QStyleOptionViewItem, model: QModelIndex):
-        res = super().sizeHint(style, model)
-        # print(res)
-        res.setHeight(self.base_height)
-        res.setWidth(max(500, res.width()))
-        return res
-
-
-class ColormapComboBox(QComboBox):
-    def __init__(self, parent):
-        super().__init__(parent)
-        view = QListView()
-        view.setMinimumWidth(COLORMAP_WIDTH + TEXT_WIDTH)
-        view.setItemDelegate(ColorStyledDelegate(ENTRY_HEIGHT))
-        self.setView(view)
